requirements.txt
Django==1.4
django-countries==1.2
django-extensions==0.9
django-tastypie==0.9.11
